b=22514 Update RHEL5.5 & OEL5.5 to latest kernel.
--RHEL5 2.6.18-194.17.1.el5.
  --OEL5  2.6.18-194.17.1.0.1.el5.
  --Switch using 'inkernel' OFED stack.
  --Build fixes for ppc64 & ia64.

+++ lustre/obdfilter/filter_io_26.c
@@ -378,8 +378,8 @@ int filter_do_bio(struct obd_export *exp, struct inode *inode,
                                        q->max_phys_segments,
                                        bio_hw_segments(q, bio),
                                        q->max_hw_segments,
-                                       (u64)bio->bi_sector,
-                                       (u64)sector);
+                                       (unsigned long long)bio->bi_sector,
+                                       (unsigned long long)sector);
 
                                 record_start_io(iobuf, rw, bio->bi_size, exp);
                                 rc = fsfilt_send_bio(rw, obd, inode, bio);
